buddy/requirements.txt
pandas >= 1.2.0
numpy >= 1.20.0
scipy >= 1.6.0
alpaca-trade-api
django>=3.2.8
djangorestframework
python-binance
TA-Lib
